Improved ndarray casting with `astype(type, copy=False)`
* Many times array is casted to the same type, but astype creates a new
  array by default. If the type does not change, `copy=False` does not
  create a new array.

+++ nnunetv2/training/data_augmentation/custom_transforms/cascade_transforms.py
@@ -121,7 +121,7 @@ class ApplyRandomBinaryOperatorTransform(AbstractTransform):
                         if not np.any(workon):
                             continue
                         # print(np.unique(workon))
-                        res = operation(workon, selem).astype(data_dict[self.key].dtype)
+                        res = operation(workon, selem).astype(data_dict[self.key].dtype, copy=False)
                         # print('ApplyRandomBinaryOperatorTransform', c, operation, np.sum(workon), np.sum(res))
                         data_dict[self.key][b, c] = res
